Replace version computation with static version
This file is modified from upstream to directly include the  version.
We have decided to do it manually so that it is easier to test a PR, as
we won't have to have a tag for that PR.
Further, this message might create a merge conflict which would inform future
version updaters to correct the version in this file.

Signed-off-by: Jimmy Brisson <theotherjimmy@gmail.com>
Signed-off-by: David Brown <david.brown@linaro.org>